* server-src/logfile.c: Do no create the log symlink for a   checkdump or fetchdump run.
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@5975 a8d146d6-cc15-0410-8900-af154a0219e0

+++ server-src/logfile.c
@@ -198,7 +198,6 @@ make_logname(
 {
     char *conf_logdir;
     char *fname = NULL;
-    char *logf;
 
     if (datestamp == NULL)
 	datestamp = g_strdup("error-00000000");
@@ -260,13 +259,16 @@ make_logname(
 	}
     }
 
-    unlink(fname);
-    logf = g_strdup(rindex(logfile,'/')+1);
-    if (symlink(logf, fname) == -1) {
-	g_debug("Can't symlink '%s' to '%s': %s", fname, logf,
-		strerror(errno));
+    if (strcmp(process, "checkdump") != 0 &&
+	strcmp(process, "fetchdump") != 0) {
+	char *logf = g_strdup(rindex(logfile,'/')+1);
+	unlink(fname);
+	if (symlink(logf, fname) == -1) {
+	    g_debug("Can't symlink '%s' to '%s': %s", fname, logf,
+		    strerror(errno));
+	}
+	amfree(logf);
     }
-    amfree(logf);
 
     amfree(fname);
     amfree(conf_logdir);
